+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.0] - 2015-11-13
+### Added
+- Tests of setFieldValidator, getFieldValidator, isEmpty, createError and parseRegExp methods
+- Tests of register method
+- Restrict validator names to match pattern [a-zA-Z]{3,}
+- Tests of maxlength validator
+- Tests of minlength validator
+- Tests of required validator
+- Tests of pattern validator
+- Tests of notpattern validator
+- Tests of maxnumber validator
+- Support of floats in maxnumber validator
+- Tests of minnumber validator
+- Support of floats in minnumber validator
+
+### Fixed
+- Not throwing errors when expression is missing or not a function
+
 ## [1.0.0] - 2015-11-11
 ### Changed
 - Implement new formlyTransformer
@@ -31,6 +49,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 0.0.1 - 2015-11-09
 
+[1.1.0]: https://github.com/wieldo/angular-formly-validator/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/wieldo/angular-formly-validator/compare/v0.0.4...v1.0.0
 [0.0.4]: https://github.com/wieldo/angular-formly-validator/compare/v0.3.0...v0.0.4
 [0.0.3]: https://github.com/wieldo/angular-formly-validator/compare/v0.2.0...v0.0.3